Extract protocol query schemas
